refactor(Socket): replace Socket object with SocketInfo struct
